requirements.txt
python-telegram-bot==13.10
psycopg2==2.9.3
matplotlib==3.5.1
